requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-pinecone>=0.2.0
langchain-text-splitters>=0.3.0
groq>=0.12.0
pinecone>=5.7.0
pinecone-client>=5.7.0
google-generativeai>=0.8.0
requests>=2.32.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.0